pyproject.toml
[tool.poetry]
name = "capstone-kb"
version = "0.1.0"
description = "CLI-first knowledge-base prototype for the CSC 675 capstone"
authors = ["Esteban Montelongo"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
click = "^8.1.7"
PyYAML = ">=6.0.1,<7.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4,<8.0"
pytest-cov = ">=4.1,<5.0"
black = ">=22.12,<23.0"

[tool.poetry.scripts]
kb = "src.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-report=xml --cov-fail-under=98"

[tool.black]
line-length = 88
target-version = ["py37"]
include = "\\.pyi?$"

[tool.coverage.report]
exclude_lines = [
	"pragma: no cover",
]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
